refactor: service locations
